+++ coloraide/contrast/wcag21.py
@@ -18,6 +18,8 @@ class WCAG21Contrast(ColorContrast):
     def contrast(self, color1: 'Color', color2: 'Color', **kwargs: Any) -> float:
         """Contrast."""
 
-        lum1 = color1.luminance()
-        lum2 = color2.luminance()
-        return (lum1 + 0.05) / (lum2 + 0.05) if (lum1 > lum2) else (lum2 + 0.05) / (lum1 + 0.05)
+        lum1 = max(0, color1.luminance())
+        lum2 = max(0, color2.luminance())
+        if (lum1 > lum2):
+            lum1, lum2 = lum2, lum1
+        return (lum2 + 0.05) / (lum1 + 0.05)